# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PLIP User Service API
  description: 유저 서비스 API 명세서
  version: v1.0.0
servers:
- url: http://localhost
  description: Generated server url
tags:
- name: Auth - Email OTP
  description: 이메일 OTP 인증 API
- name: Auth - Signup & Social Login
  description: 회원가입 및 소셜 로그인 API
- name: Users - Me
  description: 마이페이지 API
- name: Auth - Login & Token
  description: 로그인 및 JWT 토큰 API
paths:
  /api/v1/auth/signup/local:
    post:
      tags:
      - Auth - Signup & Social Login
      summary: 이메일 회원가입
      description: 이메일 인증 완료 후 회원가입을 진행합니다.
      operationId: signupLocal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalSignupRequest"
        required: true
      responses:
        "201":
          description: 회원가입 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LocalSignupResponse"
        "400":
          description: "인증 토큰 무효, 필수 약관 미동의, 닉네임 형식 오류"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 이미 가입된 이메일
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/reissue:
    post:
      tags:
      - Auth - Login & Token
      summary: 토큰 재발급
      description: 리프레시 토큰으로 액세스·리프레시 토큰을 재발급합니다 (RTR).
      operationId: reissue
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenReissueRequest"
        required: true
      responses:
        "200":
          description: 재발급 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenReissueResponse"
        "401":
          description: 리프레시 토큰 무효
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 비활성화된 계정
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/password-reset:
    post:
      tags:
      - Auth - Login & Token
      summary: 비밀번호 재설정
      description: 이메일 OTP 인증 후 비밀번호를 재설정합니다.
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
        required: true
      responses:
        "200":
          description: 재설정 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PasswordResetResponse"
        "400":
          description: "인증 토큰 무효, 기존 비밀번호와 동일"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 비활성화된 계정
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/login/social/{provider}:
    post:
      tags:
      - Auth - Signup & Social Login
      summary: 소셜 로그인
      description: 소셜 제공자의 액세스 토큰으로 로그인합니다. 기존 가입 사용자(provider+providerUserId 일치)는
        약관 없이 JWT를 발급하며 응답 newUser=false. 신규 사용자는 약관 동의 후 가입·JWT 발급(newUser=true).
        닉네임·프로필 이미지는 신규 가입 시 소셜 제공자에서 자동 수집됩니다.
      operationId: socialLogin
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialLoginRequest"
        required: true
      responses:
        "200":
          description: 로그인/가입 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SocialLoginResponse"
        "401":
          description: 소셜 인증 실패
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: 신규 사용자 — 약관 동의 필요
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/login/local:
    post:
      tags:
      - Auth - Login & Token
      summary: 로컬 로그인
      description: 이메일과 비밀번호로 로그인하고 JWT를 발급합니다.
      operationId: loginLocal
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalLoginRequest"
        required: true
      responses:
        "200":
          description: 로그인 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LocalLoginResponse"
        "401":
          description: 이메일 또는 비밀번호 불일치
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 비활성화된 계정
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/email/otp-verify:
    post:
      tags:
      - Auth - Email OTP
      summary: OTP 검증
      description: 발송된 인증번호를 검증하고 Verification Token을 반환합니다.
      operationId: verifyOtp
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailOtpVerifyRequest"
        required: true
      responses:
        "200":
          description: 인증 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EmailOtpVerifyResponse"
        "400":
          description: 인증번호 불일치 또는 만료
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/email/otp-request:
    post:
      tags:
      - Auth - Email OTP
      summary: OTP 발송 요청
      description: "입력된 이메일로 6자리 인증번호를 발송합니다. purpose=SIGNUP은 미가입 이메일만, PASSWORD_RESET은\
        \ 기존 LOCAL 계정 이메일만 발송합니다."
      operationId: requestOtp
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailOtpRequest"
        required: true
      responses:
        "200":
          description: 인증번호 발송 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EmailOtpRequestResponse"
        "429":
          description: 요청 횟수 초과
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/users/me/password:
    patch:
      tags:
      - Users - Me
      summary: 비밀번호 변경
      description: 로그인한 사용자의 비밀번호를 변경합니다. 기존 비밀번호와 동일한 비밀번호는 사용할 수 없습니다. 변경 성공 시 Redis
        Refresh Token 전량 파기 후 새 Access/Refresh JWT 발급. 기존 Access JWT는 TTL 만료까지 stateless
        유효(실무 표준). 클라이언트는 응답 JWT로 교체 권장.
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeRequest"
        required: true
      responses:
        "200":
          description: 변경 성공
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PasswordChangeResponse"
        "400":
          description: 기존 비밀번호와 동일
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패 또는 현재 비밀번호 불일치
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - bearerAuth: []
components:
  schemas:
    LocalSignupRequest:
      type: object
      description: 이메일 회원가입 요청
      properties:
        email:
          type: string
          format: email
          description: 가입 이메일
          example: user@example.com
          minLength: 1
        verificationToken:
          type: string
          description: 이메일 인증 완료 토큰
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
          minLength: 1
        password:
          type: string
          description: 비밀번호
          example: password123!
          maxLength: 64
          minLength: 8
        nickname:
          type: string
          description: 닉네임
          example: 플립이
          maxLength: 12
          minLength: 2
        termsAgreements:
          type: array
          description: 약관 동의 목록 (필수 약관이 있을 경우 필수)
          items:
            $ref: "#/components/schemas/TermAgreementRequest"
      required:
      - email
      - nickname
      - password
      - verificationToken
    TermAgreementRequest:
      type: object
      description: 약관 동의 항목
      properties:
        termId:
          type: integer
          format: int64
          description: 약관 ID
          example: 1
        agreed:
          type: boolean
          description: 동의 여부
          example: true
      required:
      - agreed
      - termId
    ErrorResponse:
      type: object
      description: 에러 응답
      properties:
        code:
          type: string
          description: 에러 코드
          example: OTP_001
        message:
          type: string
          description: 에러 메시지
          example: 요청 횟수가 제한을 초과했습니다.
    LocalSignupResponse:
      type: object
      description: 이메일 회원가입 응답
      properties:
        accessToken:
          type: string
          description: 액세스 토큰
        refreshToken:
          type: string
          description: 리프레시 토큰
        accessTokenExpiresIn:
          type: integer
          format: int64
          description: 액세스 토큰 만료 시간(초)
          example: 3600
        userUuid:
          type: string
          description: 사용자 UUID
          example: 01912345-6789-7abc-def0-123456789abc
        message:
          type: string
          description: 결과 메시지
          example: 회원가입이 완료되었습니다.
    TokenReissueRequest:
      type: object
      description: 토큰 재발급 요청
      properties:
        refreshToken:
          type: string
          description: 리프레시 토큰
          minLength: 1
      required:
      - refreshToken
    TokenReissueResponse:
      type: object
      description: 토큰 재발급 응답
      properties:
        accessToken:
          type: string
          description: 액세스 토큰
        refreshToken:
          type: string
          description: 리프레시 토큰
        accessTokenExpiresIn:
          type: integer
          format: int64
          description: 액세스 토큰 만료 시간(초)
          example: 3600
    PasswordResetRequest:
      type: object
      description: 비밀번호 재설정 요청
      properties:
        email:
          type: string
          format: email
          description: 계정 이메일
          example: user@example.com
          minLength: 1
        verificationToken:
          type: string
          description: 이메일 OTP 인증 완료 토큰
          minLength: 1
        newPassword:
          type: string
          description: 새 비밀번호
          example: newPassword123!
          maxLength: 64
          minLength: 8
      required:
      - email
      - newPassword
      - verificationToken
    PasswordResetResponse:
      type: object
      description: 비밀번호 재설정 응답
      properties:
        message:
          type: string
          description: 결과 메시지
          example: 비밀번호가 재설정되었습니다.
    SocialLoginRequest:
      type: object
      description: 소셜 로그인 요청
      properties:
        accessToken:
          type: string
          description: 소셜 제공자의 액세스 토큰
          example: ya29.a0AfH6SM...
          minLength: 1
        termsAgreements:
          type: array
          description: 약관 동의 목록 (신규 가입 시 필수)
          items:
            $ref: "#/components/schemas/TermAgreementRequest"
      required:
      - accessToken
    SocialLoginResponse:
      type: object
      description: 소셜 로그인 응답
      properties:
        accessToken:
          type: string
          description: 액세스 토큰
        refreshToken:
          type: string
          description: 리프레시 토큰
        accessTokenExpiresIn:
          type: integer
          format: int64
          description: 액세스 토큰 만료 시간(초)
          example: 3600
        userUuid:
          type: string
          description: 사용자 UUID
          example: 01912345-6789-7abc-def0-123456789abc
        newUser:
          type: boolean
          description: 신규 가입 여부
          example: false
    LocalLoginRequest:
      type: object
      description: 로컬 로그인 요청
      properties:
        email:
          type: string
          format: email
          description: 로그인 이메일
          example: user@example.com
          minLength: 1
        password:
          type: string
          description: 비밀번호
          example: password123!
          maxLength: 64
          minLength: 8
      required:
      - email
      - password
    LocalLoginResponse:
      type: object
      description: 로컬 로그인 응답
      properties:
        accessToken:
          type: string
          description: 액세스 토큰
        refreshToken:
          type: string
          description: 리프레시 토큰
        accessTokenExpiresIn:
          type: integer
          format: int64
          description: 액세스 토큰 만료 시간(초)
          example: 3600
        userUuid:
          type: string
          description: 사용자 UUID
          example: 01912345-6789-7abc-def0-123456789abc
    EmailOtpVerifyRequest:
      type: object
      description: 이메일 OTP 검증 요청
      properties:
        email:
          type: string
          format: email
          description: 인증 대상 이메일
          example: user@example.com
          minLength: 1
        otpCode:
          type: string
          description: 6자리 인증번호
          example: "123456"
          maxLength: 6
          minLength: 6
        purpose:
          type: string
          default: SIGNUP
          description: OTP 용도. otp-request와 동일한 값이어야 합니다.
          enum:
          - SIGNUP
          - PASSWORD_RESET
          example: SIGNUP
      required:
      - email
      - otpCode
    EmailOtpVerifyResponse:
      type: object
      description: 이메일 OTP 검증 응답
      properties:
        verificationToken:
          type: string
          description: 이메일 인증 완료 토큰 (회원가입 시 사용)
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
    EmailOtpRequest:
      type: object
      description: 이메일 OTP 발송 요청
      properties:
        email:
          type: string
          format: email
          description: 인증 대상 이메일
          example: user@example.com
          minLength: 1
        purpose:
          type: string
          default: SIGNUP
          description: "OTP 용도. SIGNUP=회원가입, PASSWORD_RESET=비밀번호 재설정"
          enum:
          - SIGNUP
          - PASSWORD_RESET
          example: SIGNUP
      required:
      - email
    EmailOtpRequestResponse:
      type: object
      description: 이메일 OTP 발송 응답
      properties:
        message:
          type: string
          description: 결과 메시지
          example: 인증번호가 발송되었습니다.
    PasswordChangeRequest:
      type: object
      description: 비밀번호 변경 요청
      properties:
        currentPassword:
          type: string
          description: 현재 비밀번호
          example: password123!
          minLength: 1
        newPassword:
          type: string
          description: 새 비밀번호
          example: newPassword123!
          maxLength: 64
          minLength: 8
      required:
      - currentPassword
      - newPassword
    PasswordChangeResponse:
      type: object
      description: 비밀번호 변경 응답
      properties:
        accessToken:
          type: string
          description: 액세스 토큰
        refreshToken:
          type: string
          description: 리프레시 토큰
        accessTokenExpiresIn:
          type: integer
          format: int64
          description: 액세스 토큰 만료 시간(초)
          example: 3600
        message:
          type: string
          description: 결과 메시지
          example: 비밀번호가 변경되었습니다.
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT Access Token
      scheme: bearer
      bearerFormat: JWT