# src/main/java/com/plip/user/adapter/in/web/dto/EmailOtpRequest.java
package com.plip.user.adapter.in.web.dto;

import com.plip.user.domain.model.OtpPurpose;
import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.AccessLevel;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@NoArgsConstructor(access = AccessLevel.PROTECTED)
@Schema(description = "이메일 OTP 발송 요청")
public class EmailOtpRequest {

	@NotBlank(message = "이메일은 필수입니다.")
	@Email(message = "이메일 형식이 올바르지 않습니다.")
	@Schema(description = "인증 대상 이메일", example = "user@example.com")
	private String email;

	@Schema(
			description = "OTP 용도. SIGNUP=회원가입, PASSWORD_RESET=비밀번호 재설정",
			example = "SIGNUP",
			defaultValue = "SIGNUP"
	)
	private OtpPurpose purpose = OtpPurpose.SIGNUP;
}